requirements.txt
fastapi==0.104.1
google-cloud-firestore==2.13.1
pydantic==2.5.2
pandas==2.0.3
scipy==1.10.1t
scikit-learn==1.3.2
